requirements.txt
agno
yfinance
openai
python-dotenv
streamlit
google-genai 
ddgs
requests
nsepython